+++ src/backend/marsha/core/tests/test_views_lti_video.py
@@ -456,6 +456,7 @@ class VideoLTIViewTestCase(TestCase):
                 "live_type": RAW,
                 "xmpp": {
                     "bosh_url": "https://xmpp-server.com/http-bind?token=xmpp_jwt",
+                    "converse_persistent_store": "localStorage",
                     "websocket_url": None,
                     "conference_url": f"{video.id}@conference.xmpp-server.com",
                     "jid": "conference.xmpp-server.com",