+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add a view to cancel reminder mails, link added in the footer of mails
 
+### Fixed
+
+- Allow to configure converse persistent store
+
 ## [4.0.0-beta.1] - 2022-02-15
 
 ### Added